requirements.txt
flask==2.0.1
mysql-connector-python==8.0.26
flasgger==0.9.5
python-dotenv==0.19.0
flask-cors==4.0.0
